# api-docs.json
{"openapi":"3.1.0","info":{"title":"File Flow","description":"API para gerenciamento de jobs e documentos, incluindo criacao, consulta e geracao de URL de upload.","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Jobs","description":"Endpoints para gerenciamento de jobs e documentos"}],"paths":{"/jobs/{jobId}/documents/{documentId}/upload-url":{"post":{"tags":["Jobs"],"summary":"Gerar URL de upload","description":"Gera uma URL temporaria para upload do documento.","operationId":"getGeneratedURL","parameters":[{"name":"jobId","in":"path","description":"Identificador do job","required":true,"schema":{"type":"integer","format":"int64"},"example":1},{"name":"documentId","in":"path","description":"Identificador do documento","required":true,"schema":{"type":"integer","format":"int64"},"example":10}],"responses":{"200":{"description":"URL gerada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/URLResponse"}}}},"400":{"description":"Regra de negocio violada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}},"/jobs/{jobId}/documents/create":{"post":{"tags":["Jobs"],"summary":"Criar documento no job","description":"Cria um novo documento vinculado a um job existente.","operationId":"createDocument","parameters":[{"name":"jobId","in":"path","description":"Identificador do job","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentRequest"}}},"required":true},"responses":{"200":{"description":"Documento criado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentResponse"}}}},"400":{"description":"Regra de negocio violada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}},"/jobs/create":{"post":{"tags":["Jobs"],"summary":"Criar job","description":"Cria um novo job com a quantidade total de documentos informada.","operationId":"createJob","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobRequest"}}},"required":true},"responses":{"200":{"description":"Job criado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobResponse"}}}},"401":{"description":"Nao autenticado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobResponse"}}}}},"security":[{"basicAuth":[]}]}},"/jobs/{jobId}":{"get":{"tags":["Jobs"],"summary":"Consultar job","description":"Retorna os dados de um job pelo identificador.","operationId":"getJob","parameters":[{"name":"jobId","in":"path","description":"Identificador do job","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Job encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobResponse"}}}},"400":{"description":"Regra de negocio violada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}},"/jobs/{jobId}/documents":{"get":{"tags":["Jobs"],"summary":"Listar documentos do job","description":"Retorna todos os documentos associados a um job.","operationId":"getDocumentsByJobId","parameters":[{"name":"jobId","in":"path","description":"Identificador do job","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Documentos encontrados","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ListOfDocuments"}}}},"400":{"description":"Regra de negocio violada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}},"/jobs/{jobId}/documents/{documentId}":{"get":{"tags":["Jobs"],"summary":"Consultar documento","description":"Retorna os dados de um documento especifico dentro de um job.","operationId":"getDocumentById","parameters":[{"name":"jobId","in":"path","description":"Identificador do job","required":true,"schema":{"type":"integer","format":"int64"},"example":1},{"name":"documentId","in":"path","description":"Identificador do documento","required":true,"schema":{"type":"integer","format":"int64"},"example":10}],"responses":{"200":{"description":"Documento encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentResponse"}}}},"400":{"description":"Regra de negocio violada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}}}}}},"components":{"schemas":{"URLResponse":{"type":"object","properties":{"jobId":{"type":"integer","format":"int64"},"documentId":{"type":"integer","format":"int64"},"uploadURL":{"type":"string"},"expiresInSeconds":{"type":"integer","format":"int32"}}},"ProblemDetail":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"title":{"type":"string"},"status":{"type":"integer","format":"int32"},"detail":{"type":"string"},"instance":{"type":"string","format":"uri"},"properties":{"type":"object","additionalProperties":{}}}},"DocumentRequest":{"type":"object","properties":{"originalFilename":{"type":"string"},"contentType":{"type":"string"}}},"DocumentResponse":{"type":"object","properties":{"jobId":{"type":"integer","format":"int64"},"documentId":{"type":"integer","format":"int64"},"status":{"type":"string"},"originalFilename":{"type":"string"},"contentType":{"type":"string"},"rawKey":{"type":"string"},"resultKey":{"type":"string"},"errorMessage":{"type":"string"},"sizeBytes":{"type":"integer","format":"int64"},"eTag":{"type":"string"}}},"JobRequest":{"type":"object","properties":{"totalDocuments":{"type":"integer","format":"int32"}}},"JobResponse":{"type":"object","properties":{"jobId":{"type":"integer","format":"int64"},"status":{"type":"string"},"totalDocuments":{"type":"integer","format":"int32"},"documentsCreated":{"type":"integer","format":"int32"},"doneDocuments":{"type":"integer","format":"int32"},"failedDocuments":{"type":"integer","format":"int32"}}},"ListOfDocuments":{"type":"object","properties":{"jobId":{"type":"integer","format":"int64"},"documents":{"type":"array","items":{"$ref":"#/components/schemas/DocumentResponse"}}}}}}}